+++ Sources/App/Common/ServiceRegistry/Providers/DomainServiceProvider.swift
@@ -19,10 +19,9 @@ import Vapor
 ///   - Provides intelligent caching for cost optimization
 ///
 /// ### Rules Generation Domain
-/// - **RulesOrchestrationService**: AI-powered game rules generation
-///   - Coordinates input validation, AI generation, and response processing
-///   - Handles prompt engineering and content optimization
-///   - Provides intelligent caching with TTL management
+/// - Rules generation logic moved directly to GenerateRulesUseCase for simplified architecture
+///   - Eliminates unnecessary service abstraction layer
+///   - Direct implementation in use case reduces complexity
 ///
 /// ### AI Response Validation Domain
 /// - **AIResponseValidationService**: Centralized AI response security validation
@@ -79,10 +78,8 @@ public struct DomainServiceProvider: ServiceProvider {
         
         // MARK: - Rules Generation Domain Services
         
-        /// RulesOrchestrationService for AI-powered game rules generation
-        registry.register(RulesOrchestrationService.self) { app in
-            DefaultRulesOrchestrationService()
-        }
+        // Rules generation logic has been moved directly to GenerateRulesUseCase
+        // for simplified architecture without unnecessary abstraction layers
         
         // MARK: - AI Response Validation Domain Services
