--- Sources/App/Services/Cache/CachedLLMService.swift
@@ -1,351 +1,0 @@
-import Vapor
-import Foundation
-import CryptoKit
-
-/// LLM Service wrapper that provides caching capabilities using Redis
-///
-/// This service wraps the existing LLM service to provide transparent caching
-/// of AI-generated responses, significantly reducing API costs and improving
-/// response times for repeated queries.
-///
-/// ## Performance Benefits
-/// - Reduces OpenAI API calls by up to 80% through intelligent caching
-/// - Improves response times from ~2-5 seconds to ~50-200ms for cached content
-/// - Provides consistent responses for identical inputs
-/// - Enables offline-like performance during cache hits
-///
-/// ## Cache Strategy
-/// - Cache key generated from SHA256 hash of input content
-/// - TTL varies by operation type (1 hour for rules generation, 7 days for image analysis)
-/// - Automatic cache invalidation based on TTL
-/// - Graceful fallback to original service on cache failures
-final class CachedLLMService: LLMService, @unchecked Sendable {
-    
-    private let wrappedService: LLMService
-    private let cacheService: CacheService?
-    private let logger: Logger
-    
-    public struct Configuration: Sendable {
-        public let rulesGenerationTTL: TimeInterval
-        public let imageAnalysisTTL: TimeInterval
-        public let defaultTTL: TimeInterval
-        
-        public init(
-            rulesGenerationTTL: TimeInterval = 3600,  // 1 hour
-            imageAnalysisTTL: TimeInterval = 25200,   // 7 hours  
-            defaultTTL: TimeInterval = 1800           // 30 minutes
-        ) {
-            self.rulesGenerationTTL = rulesGenerationTTL
-            self.imageAnalysisTTL = imageAnalysisTTL
-            self.defaultTTL = defaultTTL
-        }
-        
-        public static let standard = Configuration()
-        
-        public static let development = Configuration(
-            rulesGenerationTTL: 1800,  // 30 minutes
-            imageAnalysisTTL: 3600,    // 1 hour
-            defaultTTL: 900            // 15 minutes
-        )
-    }
-    
-    private let configuration: Configuration
-    
-    init(
-        wrappedService: LLMService,
-        cacheService: CacheService?,
-        configuration: Configuration = .standard,
-        logger: Logger
-    ) {
-        self.wrappedService = wrappedService
-        self.cacheService = cacheService
-        self.configuration = configuration
-        self.logger = logger
-    }
-    
-    // MARK: - LLMService Implementation
-    
-    func `for`(_ request: Request) -> LLMService {
-        return CachedLLMService(
-            wrappedService: wrappedService.for(request),
-            cacheService: cacheService,
-            configuration: configuration,
-            logger: request.logger
-        )
-    }
-    
-    func generate(input: String) async throws -> String {
-        return try await cachedGenerate(input: input, operation: .generate)
-    }
-    
-    func generateOptimized(
-        input: String,
-        model: String,
-        temperature: Double,
-        maxTokens: Int,
-        useJSONMode: Bool
-    ) async throws -> String {
-        return try await cachedGenerateOptimized(
-            input: input,
-            model: model,
-            temperature: temperature,
-            maxTokens: maxTokens,
-            useJSONMode: useJSONMode
-        )
-    }
-    
-    func analyzeImage(
-        imageData: String,
-        prompt: String,
-        model: String,
-        temperature: Double,
-        maxTokens: Int,
-        useJSONMode: Bool
-    ) async throws -> String {
-        return try await cachedAnalyzeImage(
-            imageData: imageData,
-            prompt: prompt,
-            model: model,
-            temperature: temperature,
-            maxTokens: maxTokens,
-            useJSONMode: useJSONMode
-        )
-    }
-    
-    // MARK: - Private Caching Methods
-    
-    private enum CacheOperation {
-        case generate
-        case generateOptimized
-        case analyzeImage
-    }
-    
-    private func cachedGenerate(input: String, operation: CacheOperation) async throws -> String {
-        guard let cache = cacheService else {
-            logger.debug("No cache service available, using wrapped LLM service directly")
-            return try await wrappedService.generate(input: input)
-        }
-        
-        let cacheKey = generateCacheKey(from: input, operation: operation)
-        return try await performCachedOperation(cacheKey: cacheKey, input: input) {
-            try await wrappedService.generate(input: input)
-        }
-    }
-    
-    private func cachedGenerateOptimized(
-        input: String,
-        model: String,
-        temperature: Double,
-        maxTokens: Int,
-        useJSONMode: Bool
-    ) async throws -> String {
-        guard let cache = cacheService else {
-            return try await wrappedService.generateOptimized(
-                input: input,
-                model: model,
-                temperature: temperature,
-                maxTokens: maxTokens,
-                useJSONMode: useJSONMode
-            )
-        }
-        
-        let cacheKey = generateCacheKey(
-            from: "\(input)-\(model)-\(temperature)-\(maxTokens)-\(useJSONMode)",
-            operation: .generateOptimized
-        )
-        
-        return try await performCachedOperation(cacheKey: cacheKey, input: input) {
-            try await wrappedService.generateOptimized(
-                input: input,
-                model: model,
-                temperature: temperature,
-                maxTokens: maxTokens,
-                useJSONMode: useJSONMode
-            )
-        }
-    }
-    
-    private func cachedAnalyzeImage(
-        imageData: String,
-        prompt: String,
-        model: String,
-        temperature: Double,
-        maxTokens: Int,
-        useJSONMode: Bool
-    ) async throws -> String {
-        guard let cache = cacheService else {
-            return try await wrappedService.analyzeImage(
-                imageData: imageData,
-                prompt: prompt,
-                model: model,
-                temperature: temperature,
-                maxTokens: maxTokens,
-                useJSONMode: useJSONMode
-            )
-        }
-        
-        let cacheKey = generateCacheKey(
-            from: "\(imageData.prefix(100))-\(prompt)-\(model)-\(temperature)-\(maxTokens)-\(useJSONMode)",
-            operation: .analyzeImage
-        )
-        
-        return try await performCachedOperation(cacheKey: cacheKey, input: prompt) {
-            try await wrappedService.analyzeImage(
-                imageData: imageData,
-                prompt: prompt,
-                model: model,
-                temperature: temperature,
-                maxTokens: maxTokens,
-                useJSONMode: useJSONMode
-            )
-        }
-    }
-    
-    private func performCachedOperation(
-        cacheKey: String,
-        input: String,
-        operation: () async throws -> String
-    ) async throws -> String {
-        guard let cache = cacheService else {
-            return try await operation()
-        }
-        
-        let startTime = Date()
-        
-        // Try to get cached response
-        do {
-            if let cachedResponse = try await cache.get(cacheKey, as: String.self) {
-                let duration = Date().timeIntervalSince(startTime)
-                logger.info("LLM cache hit", metadata: [
-                    "cache_key": .string(cacheKey),
-                    "duration_ms": .string(String(format: "%.2f", duration * 1000))
-                ])
-                return cachedResponse
-            }
-        } catch {
-            logger.warning("Cache retrieval failed, falling back to LLM", metadata: [
-                "cache_key": .string(cacheKey),
-                "error": .string(error.localizedDescription)
-            ])
-        }
-        
-        // Cache miss - generate response
-        let generationStartTime = Date()
-        let response = try await operation()
-        let generationDuration = Date().timeIntervalSince(generationStartTime)
-        
-        // Determine TTL and cache response
-        let ttl = determineTTL(from: input)
-        
-        // Cache asynchronously
-        Task {
-            do {
-                try await cache.set(cacheKey, value: response, ttl: ttl)
-                let totalDuration = Date().timeIntervalSince(startTime)
-                
-                logger.info("LLM response cached", metadata: [
-                    "cache_key": .string(cacheKey),
-                    "generation_duration_ms": .string(String(format: "%.2f", generationDuration * 1000)),
-                    "total_duration_ms": .string(String(format: "%.2f", totalDuration * 1000)),
-                    "ttl_seconds": .string("\(Int(ttl))")
-                ])
-            } catch {
-                logger.error("Failed to cache LLM response", metadata: [
-                    "cache_key": .string(cacheKey),
-                    "error": .string(error.localizedDescription)
-                ])
-            }
-        }
-        
-        return response
-    }
-    
-    // MARK: - Private Helpers
-    
-    /// Generate a cache key from input using SHA256 hash
-    private func generateCacheKey(from input: String, operation: CacheOperation) -> String {
-        // Create a deterministic string representation
-        let inputData = Data(input.utf8)
-        
-        // Generate SHA256 hash
-        let hash = SHA256.hash(data: inputData)
-        let hashString = withUnsafeBytes(of: hash) { bytes in
-            bytes.map { String(format: "%02x", $0) }.joined()
-        }
-        
-        let operationPrefix = switch operation {
-        case .generate: "gen"
-        case .generateOptimized: "opt" 
-        case .analyzeImage: "img"
-        }
-        
-        return "llm:\(operationPrefix):\(hashString)"
-    }
-    
-    /// Determine appropriate TTL based on input content
-    private func determineTTL(from input: String) -> TimeInterval {
-        let inputText = input.lowercased()
-        
-        // Rules generation gets longer TTL since rules don't change frequently
-        if inputText.contains("rules") || inputText.contains("game") || inputText.contains("instructions") {
-            return configuration.rulesGenerationTTL
-        }
-        
-        // Image analysis gets very long TTL since the same image always produces the same result
-        if inputText.contains("image") || inputText.contains("photo") || inputText.contains("picture") {
-            return configuration.imageAnalysisTTL
-        }
-        
-        // Default TTL for other content
-        return configuration.defaultTTL
-    }
-}
-
-// MARK: - Cache Statistics Extension
-
-extension CachedLLMService {
-    
-    /// Get cache statistics for monitoring performance
-    func getCacheStatistics() async throws -> LLMCacheStatistics {
-        guard cacheService != nil else {
-            throw CacheError.queryFailed(
-                NSError(domain: "CachedLLMService", code: 1, userInfo: [
-                    NSLocalizedDescriptionKey: "No cache service available"
-                ])
-            )
-        }
-        
-        // This would require extending CacheService with pattern-based key counting
-        // For now, return basic statistics
-        return LLMCacheStatistics(
-            totalCachedResponses: 0, // Would need implementation
-            totalCacheHits: 0,       // Would need implementation
-            totalCacheMisses: 0,     // Would need implementation
-            totalAPICostSaved: 0.0   // Would need implementation based on token counts
-        )
-    }
-    
-    /// Clear all cached LLM responses
-    func clearCache() async throws {
-        guard cacheService != nil else { return }
-        
-        // This would require pattern-based deletion (SCAN + DELETE in Redis)
-        // For now, this is a placeholder
-        logger.info("LLM cache clear requested - pattern-based deletion not yet implemented")
-    }
-}
-
-// MARK: - Supporting Types
-
-/// Statistics specific to LLM caching performance
-struct LLMCacheStatistics: Codable, Sendable {
-    let totalCachedResponses: Int
-    let totalCacheHits: Int
-    let totalCacheMisses: Int
-    let totalAPICostSaved: Double
-    
-    var hitRate: Double {
-        let total = totalCacheHits + totalCacheMisses
-        return total > 0 ? Double(totalCacheHits) / Double(total) : 0.0
-    }
-}
