--- Sources/App/Services/Cache/InMemoryAICacheService.swift
@@ -1,612 +1,0 @@
-import Foundation
-import Vapor
-
-/// Service provider extension for registering the in-memory cache implementation.
-extension Application.Service.Provider where ServiceType == AICacheServiceInterface {
-    /// Registers the in-memory AI cache service as the cache provider.
-    ///
-    /// This provider configures the application to use an in-memory cache with
-    /// the following features:
-    /// - Thread-safe operations using Swift Actor model
-    /// - LRU eviction policy when capacity is reached
-    /// - TTL-based expiration with automatic cleanup
-    /// - Comprehensive statistics and monitoring
-    /// - Graceful fallback configuration on initialization errors
-    ///
-    /// ## Performance Characteristics
-    /// - **Thread Safety**: All operations are actor-isolated for safety
-    /// - **Memory Efficiency**: Automatic cleanup of expired entries
-    /// - **Eviction**: LRU-based eviction when max capacity reached
-    /// - **Monitoring**: Built-in hit/miss ratio tracking and detailed statistics
-    static var inMemory: Self {
-        .init {
-            $0.services.aiCache.use { app in
-                do {
-                    let cacheConfig = try app.configuration.cache
-                    
-                    // Convert CacheConfig to CacheConfiguration
-                    let configuration = CacheConfiguration(
-                        maxEntries: cacheConfig.maxEntries,
-                        rulesGenerationTTL: cacheConfig.rulesGenerationTTL,
-                        imageAnalysisTTL: cacheConfig.imageAnalysisTTL,
-                        cleanupInterval: cacheConfig.cleanupInterval,
-                        enableLogging: cacheConfig.enableLogging
-                    )
-                    
-                    // Create the cache service
-                    let keyGenerator = DefaultCacheKeyGeneratorService(app: app)
-                    let cacheService = InMemoryAICacheService(
-                        configuration: configuration,
-                        logger: app.logger,
-                        keyGenerator: keyGenerator
-                    )
-                    
-                    // Start the cleanup task asynchronously
-                    Task {
-                        await cacheService.setupCleanupTask()
-                    }
-                    
-                    app.logger.info("AI Cache Service configured", metadata: [
-                        "max_entries": .string("\(configuration.maxEntries)"),
-                        "rules_ttl": .string("\(configuration.rulesGenerationTTL)s"),
-                        "image_ttl": .string("\(configuration.imageAnalysisTTL)s"),
-                        "cleanup_interval": .string("\(configuration.cleanupInterval)s"),
-                        "logging_enabled": .string("\(configuration.enableLogging)")
-                    ])
-                    
-                    return cacheService
-                    
-                } catch {
-                    app.logger.error("Failed to load cache configuration: \(error)")
-                    
-                    // Fallback configuration
-                    let fallbackConfiguration = CacheConfiguration(
-                        maxEntries: 1000,
-                        rulesGenerationTTL: 3600,
-                        imageAnalysisTTL: 1800,
-                        cleanupInterval: 600,
-                        enableLogging: true
-                    )
-                    
-                    app.logger.warning("Using fallback cache configuration", metadata: [
-                        "max_entries": .string("\(fallbackConfiguration.maxEntries)"),
-                        "rules_ttl": .string("\(fallbackConfiguration.rulesGenerationTTL)s"),
-                        "image_ttl": .string("\(fallbackConfiguration.imageAnalysisTTL)s"),
-                        "cleanup_interval": .string("\(fallbackConfiguration.cleanupInterval)s")
-                    ])
-                    
-                    let keyGenerator = DefaultCacheKeyGeneratorService(app: app)
-                    return InMemoryAICacheService(
-                        configuration: fallbackConfiguration,
-                        logger: app.logger,
-                        keyGenerator: keyGenerator
-                    )
-                }
-            }
-        }
-    }
-}
-
-/// High-performance thread-safe in-memory cache for AI responses with TTL and LRU eviction.
-///
-/// This actor-based implementation provides a complete caching solution optimized for
-/// AI-generated content with the following key features:
-///
-/// ## Core Features
-/// - **Thread Safety**: Uses Swift Actor model for safe concurrent access
-/// - **TTL Expiration**: Time-based entry expiration with configurable TTL per cache type
-/// - **LRU Eviction**: Least-Recently-Used eviction when capacity limits are reached
-/// - **Automatic Cleanup**: Background task removes expired entries periodically
-/// - **Performance Monitoring**: Real-time statistics including hit ratio and utilization
-///
-/// ## Cache Types Supported
-/// - **Rules Generation**: 24-hour TTL for game rules responses
-/// - **Image Analysis**: 7-day TTL for game box recognition results
-/// - **General Content**: Configurable TTL for other AI responses
-///
-/// ## Performance Optimizations
-/// - **Cost Reduction**: Reduces OpenAI API costs by up to 80% through intelligent caching
-/// - **Response Speed**: Sub-millisecond cache lookups for repeated queries
-/// - **Memory Management**: Efficient LRU eviction prevents unbounded memory growth
-/// - **Batch Operations**: Optimized cleanup processes minimize performance impact
-///
-/// ## Monitoring & Debugging
-/// - Hit/miss ratio tracking for cache effectiveness analysis
-/// - Detailed entry-level statistics including access patterns
-/// - Cache utilization metrics for capacity planning
-/// - Comprehensive logging for debugging and optimization
-///
-/// ## Thread Safety Guarantees
-/// All public methods are actor-isolated, ensuring:
-/// - No data races between concurrent operations
-/// - Atomic updates to cache state
-/// - Consistent view of cache statistics
-/// - Safe cleanup operations without blocking reads/writes
-actor InMemoryAICacheService: AICacheServiceInterface {
-    
-    // MARK: - Private Properties
-    
-    /// Internal storage for cache entries with TTL and metadata.
-    ///
-    /// Maps cache keys to ``CacheEntry`` objects containing:
-    /// - Cached response data
-    /// - Expiration timestamps
-    /// - Access statistics
-    /// - LRU ordering information
-    private var entries: [String: CacheEntry] = [:]
-    
-    /// LRU access tracking maintaining insertion and access order.
-    ///
-    /// This array maintains the order of cache key access, with the most
-    /// recently accessed keys at the end. Used for LRU eviction when
-    /// the cache reaches capacity limits.
-    private var accessOrder: [String] = []
-    
-    /// Cache configuration containing TTL settings and capacity limits.
-    ///
-    /// Defines:
-    /// - Maximum number of entries before eviction
-    /// - TTL values for different cache types
-    /// - Cleanup interval for expired entries
-    /// - Logging preferences
-    private let configuration: CacheConfiguration
-    
-    /// Key generator service for creating and parsing cache keys.
-    ///
-    /// Provides methods for:
-    /// - Generating consistent cache keys from input data
-    /// - Extracting cache type information from keys
-    /// - Creating hash-based keys for content deduplication
-    private let keyGenerator: CacheKeyGeneratorServiceInterface
-    
-    /// Cache hit count for performance monitoring.
-    ///
-    /// Incremented each time a valid, non-expired entry is successfully
-    /// retrieved from the cache. Used to calculate hit ratio statistics.
-    private var hitCount: Int = 0
-    
-    /// Cache miss count for performance monitoring.
-    ///
-    /// Incremented each time a cache lookup fails due to:
-    /// - Key not found in cache
-    /// - Entry has expired and was removed
-    /// Used to calculate hit ratio and cache effectiveness.
-    private var missCount: Int = 0
-    
-    /// Logger for cache operations and performance monitoring.
-    ///
-    /// Used to log:
-    /// - Cache hits and misses (when logging enabled)
-    /// - Entry additions and evictions
-    /// - Cleanup operations and statistics
-    /// - Error conditions and warnings
-    private let logger: Logger
-    
-    /// Background task for automatic cleanup of expired entries.
-    ///
-    /// Runs periodically based on ``CacheConfiguration.cleanupInterval``
-    /// to remove expired entries and maintain optimal cache performance.
-    /// Automatically cancelled during service deinitialization.
-    private var cleanupTask: Task<Void, Never>?
-    
-    // MARK: - Initialization
-    
-    /// Initializes the cache service with configuration and dependencies.
-    ///
-    /// Creates a new cache instance with the specified configuration and sets up
-    /// internal data structures. The cleanup task is initialized separately via
-    /// ``setupCleanupTask()`` to avoid blocking the initializer.
-    ///
-    /// ## Initialization Process
-    /// 1. Stores configuration and dependencies
-    /// 2. Initializes empty storage and tracking structures
-    /// 3. Logs initialization parameters (if logging enabled)
-    /// 4. Defers cleanup task setup to avoid async initialization
-    ///
-    /// - Parameters:
-    ///   - configuration: Cache configuration with TTL and capacity settings
-    ///   - logger: Logger instance for operation logging
-    ///   - keyGenerator: Service for generating and parsing cache keys
-    init(configuration: CacheConfiguration, logger: Logger, keyGenerator: CacheKeyGeneratorServiceInterface) {
-        self.configuration = configuration
-        self.logger = logger
-        self.keyGenerator = keyGenerator
-        
-        if configuration.enableLogging {
-            logger.info("AI Cache Service initialized", metadata: [
-                "max_entries": .string("\(configuration.maxEntries)"),
-                "rules_ttl": .string("\(configuration.rulesGenerationTTL)s"),
-                "image_ttl": .string("\(configuration.imageAnalysisTTL)s")
-            ])
-        }
-        
-        // Cleanup task will be initialized lazily
-    }
-    
-    /// Sets up the automatic cleanup task for expired entry removal.
-    ///
-    /// Creates a background task that runs periodically to remove expired entries
-    /// and maintain optimal cache performance. The task runs at intervals specified
-    /// by ``CacheConfiguration.cleanupInterval``.
-    ///
-    /// ## Cleanup Process
-    /// - Runs in background without blocking cache operations
-    /// - Automatically handles task cancellation during shutdown
-    /// - Gracefully handles errors and cancellation
-    /// - Prevents memory leaks from expired entries
-    ///
-    /// ## Task Lifecycle
-    /// - Started after cache initialization
-    /// - Continues until service deinitialization
-    /// - Automatically cancelled in deinit
-    ///
-    /// - Note: This method should be called once after initialization
-    func setupCleanupTask() {
-        cleanupTask = Task { [weak self] in
-            while !Task.isCancelled {
-                do {
-                    try await Task.sleep(for: .seconds(self?.configuration.cleanupInterval ?? 600))
-                    await self?.cleanupExpired()
-                } catch {
-                    // Task was cancelled, which is expected during shutdown
-                    break
-                }
-            }
-        }
-    }
-    
-    deinit {
-        cleanupTask?.cancel()
-    }
-    
-    // MARK: - AICacheServiceInterface Implementation
-    
-    nonisolated func `for`(_ request: Request) -> AICacheServiceInterface {
-        let keyGenerator = request.application.serviceCache.cacheKeyGeneratorService.for(request)
-        return Self(configuration: configuration, logger: request.logger, keyGenerator: keyGenerator)
-    }
-    
-    /// Retrieves a cached value by key with automatic expiration handling.
-    ///
-    /// This method performs the following operations:
-    /// 1. Looks up the entry in the cache storage
-    /// 2. Checks if the entry has expired
-    /// 3. Updates LRU access order and statistics
-    /// 4. Returns the cached value or nil
-    ///
-    /// ## Performance Characteristics
-    /// - **Time Complexity**: O(1) for lookup, O(n) for LRU update
-    /// - **Side Effects**: Updates access statistics and LRU ordering
-    /// - **Cleanup**: Automatically removes expired entries
-    ///
-    /// ## Statistics Tracking
-    /// - Increments hit count for successful retrievals
-    /// - Increments miss count for failures or expirations
-    /// - Updates entry access timestamp and hit count
-    ///
-    /// - Parameter key: The cache key to look up
-    /// - Returns: The cached string value if found and valid, nil otherwise
-    func get(key: String) async -> String? {
-        // Check if entry exists and is not expired
-        guard let entry = entries[key], !entry.isExpired else {
-            // Entry doesn't exist or is expired
-            if entries[key] != nil {
-                // Remove expired entry
-                entries.removeValue(forKey: key)
-                accessOrder.removeAll { $0 == key }
-                
-                if configuration.enableLogging {
-                    logger.debug("Cache entry expired and removed", metadata: ["key": .string(key)])
-                }
-            }
-            
-            missCount += 1
-            return nil
-        }
-        
-        // Update access statistics and LRU order
-        let updatedEntry = entry.recordAccess()
-        entries[key] = updatedEntry
-        updateAccessOrder(for: key)
-        
-        hitCount += 1
-        
-        if configuration.enableLogging {
-            logger.debug("Cache hit", metadata: [
-                "key": .string(key),
-                "hit_count": .string("\(updatedEntry.hitCount)"),
-                "age": .string(String(format: "%.1fs", updatedEntry.age))
-            ])
-        }
-        
-        return entry.value
-    }
-    
-    /// Stores a value in the cache with the specified time-to-live.
-    ///
-    /// This method handles cache storage with automatic capacity management:
-    /// 1. Creates a new cache entry with TTL
-    /// 2. Triggers LRU eviction if at capacity
-    /// 3. Stores the entry and updates access order
-    /// 4. Logs storage operation (if enabled)
-    ///
-    /// ## Capacity Management
-    /// - Checks if cache is at maximum capacity
-    /// - Triggers LRU eviction before adding new entries
-    /// - Preserves existing entries when updating values
-    ///
-    /// ## Performance Considerations
-    /// - **Time Complexity**: O(1) for storage, O(k) for eviction where k is eviction batch size
-    /// - **Memory Usage**: Automatically manages memory through eviction
-    /// - **Concurrency**: Thread-safe through actor isolation
-    ///
-    /// - Parameters:
-    ///   - key: The cache key for the stored value
-    ///   - value: The string value to cache
-    ///   - ttl: Time-to-live in seconds for the cache entry
-    func set(key: String, value: String, ttl: TimeInterval) async {
-        let entry = CacheEntry(key: key, value: value, ttl: ttl)
-        
-        // Check if we need to evict entries to make room
-        if entries.count >= configuration.maxEntries && entries[key] == nil {
-            await evictLeastRecentlyUsed()
-        }
-        
-        // Store the entry
-        entries[key] = entry
-        updateAccessOrder(for: key)
-        
-        if configuration.enableLogging {
-            logger.debug("Cache entry stored", metadata: [
-                "key": .string(key),
-                "ttl": .string("\(ttl)s"),
-                "value_size": .string("\(value.count) chars"),
-                "total_entries": .string("\(entries.count)")
-            ])
-        }
-    }
-    
-    func exists(key: String) async -> Bool {
-        guard let entry = entries[key] else { return false }
-        
-        if entry.isExpired {
-            entries.removeValue(forKey: key)
-            accessOrder.removeAll { $0 == key }
-            return false
-        }
-        
-        return true
-    }
-    
-    func remove(key: String) async {
-        if entries.removeValue(forKey: key) != nil {
-            accessOrder.removeAll { $0 == key }
-            
-            if configuration.enableLogging {
-                logger.debug("Cache entry removed", metadata: ["key": .string(key)])
-            }
-        }
-    }
-    
-    func clear() async {
-        let previousCount = entries.count
-        entries.removeAll()
-        accessOrder.removeAll()
-        
-        if configuration.enableLogging {
-            logger.info("Cache cleared", metadata: ["removed_entries": .string("\(previousCount)")])
-        }
-    }
-    
-    func getStatistics() async -> CacheStatistics {
-        return CacheStatistics(
-            hits: hitCount,
-            misses: missCount,
-            entryCount: entries.count,
-            maxEntries: configuration.maxEntries
-        )
-    }
-    
-    /// Removes all expired entries from the cache.
-    ///
-    /// This method performs a comprehensive cleanup of expired entries:
-    /// 1. Scans all cache entries for expiration
-    /// 2. Identifies expired keys for batch removal
-    /// 3. Removes expired entries from storage and LRU tracking
-    /// 4. Logs cleanup statistics (if enabled)
-    ///
-    /// ## Cleanup Process
-    /// - **Batch Processing**: Collects expired keys before removal for efficiency
-    /// - **Atomic Updates**: Removes all expired entries in a single operation
-    /// - **LRU Maintenance**: Updates access order tracking consistently
-    /// - **Statistics**: Tracks cleanup effectiveness for monitoring
-    ///
-    /// ## Performance Characteristics
-    /// - **Time Complexity**: O(n) where n is the number of cache entries
-    /// - **Memory Recovery**: Immediately frees memory from expired entries
-    /// - **Non-blocking**: Designed to run efficiently in background
-    ///
-    /// This method is called automatically by the cleanup task and can also
-    /// be invoked manually for immediate cleanup.
-    func cleanupExpired() async {
-        let initialCount = entries.count
-        let now = Date()
-        
-        var expiredKeys: [String] = []
-        
-        for (key, entry) in entries {
-            if now > entry.expiresAt {
-                expiredKeys.append(key)
-            }
-        }
-        
-        for key in expiredKeys {
-            entries.removeValue(forKey: key)
-            accessOrder.removeAll { $0 == key }
-        }
-        
-        let removedCount = expiredKeys.count
-        
-        if removedCount > 0 && configuration.enableLogging {
-            logger.debug("Expired entries cleaned up", metadata: [
-                "removed_count": .string("\(removedCount)"),
-                "remaining_count": .string("\(entries.count)"),
-                "initial_count": .string("\(initialCount)")
-            ])
-        }
-    }
-    
-    func count() async -> Int {
-        return entries.count
-    }
-    
-    // MARK: - Private Helper Methods
-    
-    /// Updates the access order for LRU tracking and eviction decisions.
-    ///
-    /// This method maintains the LRU ordering by moving the accessed key to the
-    /// end of the access order array, marking it as the most recently used entry.
-    ///
-    /// ## LRU Algorithm Implementation
-    /// - **Removal**: Removes key from current position in access order
-    /// - **Insertion**: Adds key to end of array (most recently used)
-    /// - **Ordering**: Maintains chronological access order for eviction
-    ///
-    /// ## Performance Impact
-    /// - **Time Complexity**: O(n) due to array removal operation
-    /// - **Space Complexity**: O(1) additional space required
-    /// - **Frequency**: Called on every cache access (get/set)
-    ///
-    /// - Parameter key: The cache key that was accessed or updated
-    private func updateAccessOrder(for key: String) {
-        // Remove key from current position
-        accessOrder.removeAll { $0 == key }
-        // Add to end (most recently used)
-        accessOrder.append(key)
-    }
-    
-    /// Evicts the least recently used entries when cache reaches capacity.
-    ///
-    /// This method implements the LRU eviction policy to maintain cache capacity:
-    /// 1. Calculates the number of entries to evict (10% of max capacity)
-    /// 2. Identifies the least recently used entries from access order
-    /// 3. Removes selected entries from both storage and tracking structures
-    /// 4. Logs eviction statistics for monitoring
-    ///
-    /// ## Eviction Strategy
-    /// - **Batch Size**: Evicts 10% of max capacity or minimum 1 entry
-    /// - **Selection**: Uses LRU order (oldest accessed entries first)
-    /// - **Efficiency**: Batch eviction reduces frequent eviction overhead
-    ///
-    /// ## Performance Characteristics
-    /// - **Time Complexity**: O(k) where k is the number of entries evicted
-    /// - **Memory Recovery**: Immediately frees memory from evicted entries
-    /// - **Preservation**: Keeps most frequently accessed entries in cache
-    ///
-    /// This method is called automatically when the cache reaches its maximum
-    /// capacity and a new entry needs to be stored.
-    private func evictLeastRecentlyUsed() async {
-        guard !accessOrder.isEmpty else { return }
-        
-        // Calculate how many entries to evict (10% of max capacity or at least 1)
-        let entriesToEvict = max(1, configuration.maxEntries / 10)
-        let keysToEvict = Array(accessOrder.prefix(entriesToEvict))
-        
-        for key in keysToEvict {
-            entries.removeValue(forKey: key)
-        }
-        
-        // Remove evicted keys from access order
-        accessOrder.removeFirst(keysToEvict.count)
-        
-        if configuration.enableLogging {
-            logger.debug("LRU eviction performed", metadata: [
-                "evicted_count": .string("\(keysToEvict.count)"),
-                "remaining_count": .string("\(entries.count)")
-            ])
-        }
-    }
-    
-    // MARK: - Additional Utility Methods
-    
-    /// Returns comprehensive cache information for debugging and monitoring.
-    ///
-    /// This method provides detailed insights into cache performance and state:
-    /// - Overall cache statistics (hits, misses, utilization)
-    /// - Configuration settings and limits
-    /// - Individual entry details with access patterns
-    /// - Sorted entry list by access recency
-    ///
-    /// ## Information Provided
-    /// - **Statistics**: Hit ratio, entry count, utilization percentage
-    /// - **Configuration**: TTL settings, capacity limits, logging status
-    /// - **Entries**: Per-entry age, hit count, expiration status
-    /// - **Access Patterns**: Last accessed times and frequency data
-    ///
-    /// ## Use Cases
-    /// - Performance analysis and optimization
-    /// - Debugging cache behavior and effectiveness
-    /// - Monitoring cache utilization patterns
-    /// - Troubleshooting cache-related issues
-    ///
-    /// - Returns: Dictionary containing comprehensive cache information
-    func getDetailedInfo() async -> [String: Any] {
-        let stats = await getStatistics()
-        
-        var entryDetails: [[String: Any]] = []
-        let now = Date()
-        
-        for entry in entries.values {
-            entryDetails.append([
-                "key": entry.key,
-                "age": entry.age,
-                "ttl_remaining": entry.timeToLive,
-                "hit_count": entry.hitCount,
-                "last_accessed": entry.lastAccessedAt.timeIntervalSince(now),
-                "expired": entry.isExpired
-            ])
-        }
-        
-        // Sort by most recently accessed
-        entryDetails.sort { lhs, rhs in
-            let lhsAccessed = lhs["last_accessed"] as? TimeInterval ?? 0
-            let rhsAccessed = rhs["last_accessed"] as? TimeInterval ?? 0
-            return lhsAccessed > rhsAccessed
-        }
-        
-        return [
-            "statistics": [
-                "hits": stats.hits,
-                "misses": stats.misses,
-                "hit_ratio": stats.hitRatio,
-                "entry_count": stats.entryCount,
-                "max_entries": stats.maxEntries,
-                "utilization": stats.utilization
-            ],
-            "configuration": [
-                "max_entries": configuration.maxEntries,
-                "rules_ttl": configuration.rulesGenerationTTL,
-                "image_ttl": configuration.imageAnalysisTTL,
-                "cleanup_interval": configuration.cleanupInterval,
-                "logging_enabled": configuration.enableLogging
-            ],
-            "entries": entryDetails
-        ]
-    }
-    
-    /// Returns cache entries grouped by type
-    func getEntriesByType() async -> [AICacheType: [String]] {
-        var result: [AICacheType: [String]] = [:]
-        
-        for key in entries.keys {
-            if let type = keyGenerator.extractCacheType(from: key) {
-                if result[type] == nil {
-                    result[type] = []
-                }
-                result[type]?.append(key)
-            }
-        }
-        
-        return result
-    }
-}
