--- Tests/AppTests/Services/Cache/InMemoryAICacheServiceTests.swift
@@ -1,397 +1,0 @@
-@testable import App
-import XCTest
-import Vapor
-
-final class InMemoryAICacheServiceTests: XCTestCase {
-    var app: Application!
-    var cacheService: InMemoryAICacheService!
-    var keyGenerator: DefaultCacheKeyGeneratorService!
-    var testConfiguration: CacheConfiguration!
-    
-    override func setUpWithError() throws {
-        app = try TestWorld.makeTestAppSync()
-        
-        // Create test configuration with shorter TTLs for faster testing
-        testConfiguration = CacheConfiguration(
-            maxEntries: 10,
-            rulesGenerationTTL: 1.0,      // 1 second for fast expiration testing
-            imageAnalysisTTL: 2.0,        // 2 seconds
-            cleanupInterval: 0.5,         // 0.5 seconds for frequent cleanup
-            enableLogging: false          // Disable logging in tests
-        )
-        
-        keyGenerator = DefaultCacheKeyGeneratorService(app: app)
-        cacheService = InMemoryAICacheService(
-            configuration: testConfiguration,
-            logger: app.logger,
-            keyGenerator: keyGenerator
-        )
-    }
-    
-    override func tearDown() {
-        app.shutdown()
-    }
-    
-    // MARK: - Basic Cache Operations
-    
-    func testSetAndGet() async throws {
-        let key = "test_key"
-        let value = "test_value"
-        
-        // Initially should be nil
-        let initialValue = await cacheService.get(key: key)
-        XCTAssertNil(initialValue)
-        
-        // Set the value
-        await cacheService.set(key: key, value: value, ttl: 60.0)
-        
-        // Should retrieve the value
-        let retrievedValue = await cacheService.get(key: key)
-        XCTAssertEqual(retrievedValue, value)
-    }
-    
-    func testExists() async throws {
-        let key = "exists_test"
-        let value = "test_value"
-        
-        // Initially should not exist
-        let initialExists = await cacheService.exists(key: key)
-        XCTAssertFalse(initialExists)
-        
-        // Set the value
-        await cacheService.set(key: key, value: value, ttl: 60.0)
-        
-        // Should now exist
-        let nowExists = await cacheService.exists(key: key)
-        XCTAssertTrue(nowExists)
-    }
-    
-    func testRemove() async throws {
-        let key = "remove_test"
-        let value = "test_value"
-        
-        // Set and verify
-        await cacheService.set(key: key, value: value, ttl: 60.0)
-        var retrievedValue = await cacheService.get(key: key)
-        XCTAssertEqual(retrievedValue, value)
-        
-        // Remove and verify
-        await cacheService.remove(key: key)
-        retrievedValue = await cacheService.get(key: key)
-        XCTAssertNil(retrievedValue)
-        
-        // Should not exist
-        let exists = await cacheService.exists(key: key)
-        XCTAssertFalse(exists)
-    }
-    
-    func testClear() async throws {
-        // Set multiple values
-        for i in 0..<5 {
-            await cacheService.set(key: "key_\(i)", value: "value_\(i)", ttl: 60.0)
-        }
-        
-        // Verify they exist
-        let countBefore = await cacheService.count()
-        XCTAssertEqual(countBefore, 5)
-        
-        // Clear all
-        await cacheService.clear()
-        
-        // Verify all are gone
-        let countAfter = await cacheService.count()
-        XCTAssertEqual(countAfter, 0)
-        
-        for i in 0..<5 {
-            let value = await cacheService.get(key: "key_\(i)")
-            XCTAssertNil(value)
-        }
-    }
-    
-    func testCount() async throws {
-        // Initially empty
-        let initialCount = await cacheService.count()
-        XCTAssertEqual(initialCount, 0)
-        
-        // Add some entries
-        for i in 0..<3 {
-            await cacheService.set(key: "count_key_\(i)", value: "value_\(i)", ttl: 60.0)
-        }
-        
-        let afterAddingCount = await cacheService.count()
-        XCTAssertEqual(afterAddingCount, 3)
-        
-        // Remove one
-        await cacheService.remove(key: "count_key_1")
-        
-        let afterRemovingCount = await cacheService.count()
-        XCTAssertEqual(afterRemovingCount, 2)
-    }
-    
-    // MARK: - TTL and Expiration Tests
-    
-    func testTTLExpiration() async throws {
-        let key = "ttl_test"
-        let value = "test_value"
-        let shortTTL: TimeInterval = 0.1 // 100ms
-        
-        // Set with short TTL
-        await cacheService.set(key: key, value: value, ttl: shortTTL)
-        
-        // Should exist immediately
-        var retrievedValue = await cacheService.get(key: key)
-        XCTAssertEqual(retrievedValue, value)
-        
-        // Wait for expiration
-        try await Task.sleep(nanoseconds: 200_000_000) // 200ms
-        
-        // Should be expired now
-        retrievedValue = await cacheService.get(key: key)
-        XCTAssertNil(retrievedValue)
-        
-        let exists = await cacheService.exists(key: key)
-        XCTAssertFalse(exists)
-    }
-    
-    func testAutomaticCleanupOfExpiredEntries() async throws {
-        let key1 = "cleanup_key_1"
-        let key2 = "cleanup_key_2"
-        let value = "test_value"
-        let shortTTL: TimeInterval = 0.1 // 100ms
-        let longTTL: TimeInterval = 60.0 // 60 seconds
-        
-        // Set one with short TTL and one with long TTL
-        await cacheService.set(key: key1, value: value, ttl: shortTTL)
-        await cacheService.set(key: key2, value: value, ttl: longTTL)
-        
-        // Both should exist initially
-        let initialCount = await cacheService.count()
-        XCTAssertEqual(initialCount, 2)
-        
-        // Wait for first to expire
-        try await Task.sleep(nanoseconds: 200_000_000) // 200ms
-        
-        // Trigger cleanup
-        await cacheService.cleanupExpired()
-        
-        // Count should be reduced
-        let countAfterCleanup = await cacheService.count()
-        XCTAssertEqual(countAfterCleanup, 1)
-        
-        // Only the long TTL key should remain
-        let value1 = await cacheService.get(key: key1)
-        let value2 = await cacheService.get(key: key2)
-        XCTAssertNil(value1)
-        XCTAssertEqual(value2, value)
-    }
-    
-    // MARK: - Capacity and LRU Tests
-    
-    func testCapacityLimit() async throws {
-        let maxEntries = testConfiguration.maxEntries
-        
-        // Fill cache to capacity
-        for i in 0..<maxEntries {
-            await cacheService.set(key: "capacity_key_\(i)", value: "value_\(i)", ttl: 60.0)
-        }
-        
-        let countAtCapacity = await cacheService.count()
-        XCTAssertEqual(countAtCapacity, maxEntries)
-        
-        // Add one more entry - should trigger LRU eviction
-        await cacheService.set(key: "overflow_key", value: "overflow_value", ttl: 60.0)
-        
-        // Count should still be at max
-        let countAfterOverflow = await cacheService.count()
-        XCTAssertEqual(countAfterOverflow, maxEntries)
-        
-        // The new key should exist
-        let overflowValue = await cacheService.get(key: "overflow_key")
-        XCTAssertEqual(overflowValue, "overflow_value")
-    }
-    
-    func testLRUEviction() async throws {
-        // Fill cache to capacity
-        for i in 0..<testConfiguration.maxEntries {
-            await cacheService.set(key: "lru_key_\(i)", value: "value_\(i)", ttl: 60.0)
-        }
-        
-        // Access the first key to make it recently used
-        let firstValue = await cacheService.get(key: "lru_key_0")
-        XCTAssertEqual(firstValue, "value_0")
-        
-        // Add a new entry to trigger eviction
-        await cacheService.set(key: "new_key", value: "new_value", ttl: 60.0)
-        
-        // The first key should still exist (was recently accessed)
-        let stillExists = await cacheService.get(key: "lru_key_0")
-        XCTAssertEqual(stillExists, "value_0")
-        
-        // Some other key should have been evicted (likely lru_key_1 since it wasn't accessed)
-        let _ = await cacheService.get(key: "lru_key_1")
-        // Note: This test is probabilistic and depends on LRU implementation details
-        // We mainly check that capacity is maintained and some eviction occurred
-        let finalCount = await cacheService.count()
-        XCTAssertEqual(finalCount, testConfiguration.maxEntries)
-    }
-    
-    // MARK: - Statistics Tests
-    
-    func testStatistics() async throws {
-        // Get initial statistics
-        var stats = await cacheService.getStatistics()
-        XCTAssertEqual(stats.hits, 0)
-        XCTAssertEqual(stats.misses, 0)
-        XCTAssertEqual(stats.entryCount, 0)
-        XCTAssertEqual(stats.hitRatio, 0.0)
-        
-        // Set some values
-        await cacheService.set(key: "stats_key_1", value: "value_1", ttl: 60.0)
-        await cacheService.set(key: "stats_key_2", value: "value_2", ttl: 60.0)
-        
-        // Get hits and misses
-        let hit1 = await cacheService.get(key: "stats_key_1")
-        XCTAssertNotNil(hit1) // Hit
-        
-        let miss1 = await cacheService.get(key: "nonexistent_key")
-        XCTAssertNil(miss1) // Miss
-        
-        let hit2 = await cacheService.get(key: "stats_key_2")
-        XCTAssertNotNil(hit2) // Hit
-        
-        // Check updated statistics
-        stats = await cacheService.getStatistics()
-        XCTAssertEqual(stats.hits, 2)
-        XCTAssertEqual(stats.misses, 1)
-        XCTAssertEqual(stats.entryCount, 2)
-        XCTAssertEqual(stats.hitRatio, 2.0/3.0 * 100.0, accuracy: 0.001)
-    }
-    
-    func testStatisticsWithEviction() async throws {
-        // Fill cache beyond capacity to trigger evictions
-        let maxEntries = testConfiguration.maxEntries
-        
-        for i in 0..<(maxEntries + 2) {
-            await cacheService.set(key: "eviction_key_\(i)", value: "value_\(i)", ttl: 60.0)
-        }
-        
-        let stats = await cacheService.getStatistics()
-        XCTAssertEqual(stats.entryCount, maxEntries)
-        // Note: evictionCount is not available in CacheStatistics
-    }
-    
-    // MARK: - Memory Usage Tests
-    
-    func testMemoryUsageCalculation() async throws {
-        // Add some entries
-        await cacheService.set(key: "memory_key_1", value: "small_value", ttl: 60.0)
-        await cacheService.set(key: "memory_key_2", value: "larger_value_with_more_content", ttl: 60.0)
-        
-        let stats = await cacheService.getStatistics()
-        // Note: memoryUsage is not available in CacheStatistics
-        // Test that we can get valid statistics instead
-        XCTAssertEqual(stats.entryCount, 2)
-        
-        // Add another entry and verify count increases
-        await cacheService.set(key: "memory_key_3", value: String(repeating: "x", count: 100), ttl: 60.0)
-        
-        let newStats = await cacheService.getStatistics()
-        XCTAssertEqual(newStats.entryCount, 3)
-    }
-    
-    // MARK: - Service Pattern Tests
-    
-    func testServicePatternForMethod() throws {
-        let request = Request(
-            application: app,
-            method: .GET,
-            url: "http://localhost/test",
-            on: app.eventLoopGroup.next()
-        )
-        
-        let serviceForRequest = cacheService.for(request)
-        XCTAssertTrue(serviceForRequest is InMemoryAICacheService)
-        
-        // Should return a new instance for request-specific operations
-        // Note: Cannot use identity comparison on protocol types, test type instead
-        XCTAssertTrue(serviceForRequest is InMemoryAICacheService)
-    }
-    
-    // MARK: - Concurrent Access Tests
-    
-    func testConcurrentAccess() async throws {
-        let key = "concurrent_key"
-        let numberOfTasks = 10
-        
-        // Create multiple concurrent tasks that set/get values
-        await withTaskGroup(of: Void.self) { group in
-            for i in 0..<numberOfTasks {
-                group.addTask { [cacheService = self.cacheService!] in
-                    await cacheService.set(key: "\(key)_\(i)", value: "value_\(i)", ttl: 60.0)
-                    let _ = await cacheService.get(key: "\(key)_\(i)")
-                }
-            }
-        }
-        
-        // All entries should be stored
-        let finalCount = await cacheService.count()
-        XCTAssertEqual(finalCount, numberOfTasks)
-        
-        // Verify statistics are consistent
-        let stats = await cacheService.getStatistics()
-        XCTAssertEqual(stats.entryCount, numberOfTasks)
-        XCTAssertEqual(stats.hits, numberOfTasks) // Each task did one get
-    }
-    
-    // MARK: - Edge Cases
-    
-    func testEmptyKeyAndValue() async throws {
-        let emptyKey = ""
-        let emptyValue = ""
-        
-        // Should handle empty strings
-        await cacheService.set(key: emptyKey, value: emptyValue, ttl: 60.0)
-        let retrievedValue = await cacheService.get(key: emptyKey)
-        XCTAssertEqual(retrievedValue, emptyValue)
-    }
-    
-    func testZeroTTL() async throws {
-        let key = "zero_ttl_key"
-        let value = "test_value"
-        
-        // Set with zero TTL (should expire immediately)
-        await cacheService.set(key: key, value: value, ttl: 0.0)
-        
-        // Should not be retrievable
-        let retrievedValue = await cacheService.get(key: key)
-        XCTAssertNil(retrievedValue)
-    }
-    
-    func testNegativeTTL() async throws {
-        let key = "negative_ttl_key"
-        let value = "test_value"
-        
-        // Set with negative TTL (should expire immediately)
-        await cacheService.set(key: key, value: value, ttl: -1.0)
-        
-        // Should not be retrievable
-        let retrievedValue = await cacheService.get(key: key)
-        XCTAssertNil(retrievedValue)
-    }
-    
-    func testLargeValue() async throws {
-        let key = "large_value_key"
-        let largeValue = String(repeating: "A", count: 10000) // 10KB string
-        
-        // Should handle large values
-        await cacheService.set(key: key, value: largeValue, ttl: 60.0)
-        let retrievedValue = await cacheService.get(key: key)
-        XCTAssertEqual(retrievedValue, largeValue)
-        
-        // Memory usage should reflect the large value
-        let stats = await cacheService.getStatistics()
-        // Note: memoryUsage is not available in CacheStatistics
-        // Verify the entry was stored
-        XCTAssertEqual(stats.entryCount, 1)
-    }
-}
