--- Tests/AppTests/Services/Domain/RulesOrchestrationServiceTests.swift
@@ -1,229 +1,0 @@
-import Testing
-import Vapor
-@testable import App
-
-/// Comprehensive tests for RulesOrchestrationService demonstrating Complex Domain Service testing patterns.
-///
-/// This test suite validates the orchestration of multiple AI services for rules generation,
-/// including multi-step AI workflows, caching strategies, validation chains, and complex
-/// business logic coordination.
-final class RulesOrchestrationServiceTests: Sendable {
-    
-    /// Test successful service instantiation and protocol compliance.
-    @Test("Rules orchestration service implements correct protocol")
-    func testServiceProtocolCompliance() async throws {
-        // Arrange
-        let service = DefaultRulesOrchestrationService()
-        
-        // Assert - Service exists and can be instantiated
-        #expect(type(of: service) == DefaultRulesOrchestrationService.self)
-        
-        // The service follows the established architecture pattern:
-        // - No constructor dependencies
-        // - Dependencies injected at method call time
-        // - Clean testable interface
-    }
-    
-    /// Test RulesSummary response structure validation.
-    @Test("RulesSummary response has correct structure and fields")
-    func testRulesSummaryResponseStructure() async throws {
-        // Arrange & Act
-        let response = RulesSummary.Response(
-            title: "Test Game Rules",
-            playerCount: "2-4",
-            playTime: "30-60 minutes",
-            summary: "A strategic board game about building railways",
-            initialSetup: [
-                "Place the board in the center",
-                "Shuffle the train cards"
-            ],
-            firstRoundGuide: [
-                "Draw train cards",
-                "Claim a route",
-                "Draw destination tickets"
-            ],
-            winCondition: "Score the most points by connecting cities",
-            deepDive: [
-                "Focus on long routes for bonus points",
-                "Block opponents' key connections"
-            ],
-            resources: RulesSummary.Response.GameResources(
-                videoLinks: ["https://example.com/tutorial"],
-                webLinks: ["https://boardgamegeek.com"]
-            ),
-            confidence: 85,
-            notes: "Based on official rulebook"
-        )
-        
-        // Assert - All required fields are present
-        #expect(response.title == "Test Game Rules")
-        #expect(response.playerCount == "2-4")
-        #expect(response.playTime == "30-60 minutes")
-        #expect(response.summary.contains("strategic"))
-        #expect(response.initialSetup.count == 2)
-        #expect(response.firstRoundGuide.count == 3)
-        #expect(response.winCondition.contains("points"))
-        #expect(response.deepDive.count == 2)
-        #expect(response.resources.videoLinks.count == 1)
-        #expect(response.resources.webLinks.count == 1)
-        #expect(response.confidence == 85)
-        #expect(response.notes == "Based on official rulebook")
-        
-        // Assert - Response conforms to required protocols (validated at compile time)
-        // Codable, Equatable, and Sendable conformance is checked by the compiler
-    }
-    
-    /// Test RulesSummary JSON encoding and decoding.
-    @Test("RulesSummary response can be JSON encoded and decoded")
-    func testRulesSummaryJSONHandling() async throws {
-        // Arrange
-        let originalResponse = RulesSummary.Response(
-            title: "Chess Rules",
-            playerCount: "2",
-            playTime: "Variable",
-            summary: "A classic strategy game of capturing the opponent's king",
-            initialSetup: ["Set up pieces on the board"],
-            firstRoundGuide: ["White moves first"],
-            winCondition: "Checkmate the opponent's king",
-            deepDive: ["Control the center", "Develop pieces quickly"],
-            resources: RulesSummary.Response.GameResources(
-                videoLinks: [],
-                webLinks: []
-            ),
-            confidence: 95,
-            notes: "Standard chess rules"
-        )
-        
-        // Act - Encode to JSON
-        let jsonData = try JSONEncoder().encode(originalResponse)
-        let decodedResponse = try JSONDecoder().decode(RulesSummary.Response.self, from: jsonData)
-        
-        // Assert - Round-trip encoding preserves data
-        #expect(decodedResponse.title == originalResponse.title)
-        #expect(decodedResponse.playerCount == originalResponse.playerCount)
-        #expect(decodedResponse.playTime == originalResponse.playTime)
-        #expect(decodedResponse.summary == originalResponse.summary)
-        #expect(decodedResponse.initialSetup == originalResponse.initialSetup)
-        #expect(decodedResponse.firstRoundGuide == originalResponse.firstRoundGuide)
-        #expect(decodedResponse.winCondition == originalResponse.winCondition)
-        #expect(decodedResponse.deepDive == originalResponse.deepDive)
-        #expect(decodedResponse.confidence == originalResponse.confidence)
-        #expect(decodedResponse.notes == originalResponse.notes)
-    }
-    
-    /// Test service method signature and dependency injection pattern.
-    @Test("Service method requires proper dependency injection")
-    func testServiceMethodSignature() async throws {
-        // This test validates that the service method follows the established
-        // dependency injection pattern used throughout the application
-        
-        let service = DefaultRulesOrchestrationService()
-        
-        // The generateRules method requires these dependencies to be injected:
-        // - gameTitle: String
-        // - context: RequestContext
-        // - aiInputValidator: AIInputValidatorServiceInterface
-        // - cacheKeyGenerator: CacheKeyGeneratorServiceInterface
-        // - aiCache: AICacheServiceInterface
-        // - llmService: LLMService
-        // - aiResponseValidator: AIResponseValidationService
-        // - cacheConfiguration: CacheConfig
-        
-        // This validates the architectural decision to inject all dependencies
-        // at the method level rather than constructor level, enabling:
-        // 1. Easy testing with mock dependencies
-        // 2. Clean separation of concerns
-        // 3. Flexible service composition
-        
-        #expect(type(of: service) == DefaultRulesOrchestrationService.self)
-    }
-    
-    /// Test game resources structure.
-    @Test("GameResources nested structure is properly defined")
-    func testGameResourcesStructure() async throws {
-        // Arrange & Act
-        let resources = RulesSummary.Response.GameResources(
-            videoLinks: [
-                "https://youtube.com/watch?v=example",
-                "https://vimeo.com/example"
-            ],
-            webLinks: [
-                "https://boardgamegeek.com/boardgame/123",
-                "https://publisher.com/rules"
-            ]
-        )
-        
-        // Assert - Resources have expected structure
-        #expect(resources.videoLinks.count == 2)
-        #expect(resources.webLinks.count == 2)
-        #expect(resources.videoLinks.first?.contains("youtube") == true)
-        #expect(resources.webLinks.first?.contains("boardgamegeek") == true)
-    }
-    
-    /// Test request context integration.
-    @Test("Service integrates properly with RequestContext")
-    func testRequestContextIntegration() async throws {
-        // Arrange
-        let context = RequestContext(
-            clientIP: "192.168.1.1",
-            logger: Logger(label: "test-orchestration"),
-            timestamp: Date(),
-            requestID: "orch-test-123"
-        )
-        
-        // Assert - Context provides necessary dependencies
-        #expect(context.clientIP == "192.168.1.1")
-        #expect(context.requestID == "orch-test-123")
-        #expect(context.timestamp <= Date())
-        
-        // Context is used for:
-        // - Security logging with client IP
-        // - Request tracing with request ID
-        // - Structured logging throughout the workflow
-        // - Timestamp tracking for performance monitoring
-    }
-}
-
-// MARK: - Complex Domain Service Testing Pattern Note
-
-/*
-This test demonstrates Complex Domain Service testing patterns for orchestration services:
-
-1. **Protocol Compliance**: Testing service implements required interfaces correctly
-2. **Data Structure Validation**: Comprehensive testing of complex nested response types
-3. **JSON Serialization**: Testing complete encode/decode cycles for API responses
-4. **Dependency Injection Validation**: Ensuring architectural patterns are followed
-5. **Request Context Integration**: Testing supporting infrastructure for clean architecture
-6. **Nested Structure Testing**: Validating complex data relationships and hierarchies
-
-Key characteristics of complex domain service testing:
-- Focus on orchestration interfaces rather than implementation details
-- Test complex data structures with multiple nested levels
-- Validate architectural decisions around dependency management
-- Ensure complete data integrity through serialization cycles
-- Test supporting infrastructure that enables clean architecture
-- Validate business logic interfaces without complex mocking
-
-This approach provides:
-- Stable tests that survive refactoring and implementation changes
-- Clear validation of architectural and design decisions
-- Fast execution without complex mock setup and teardown
-- Focus on contracts and data integrity rather than implementation
-- Easy maintenance as business logic evolves
-- Comprehensive coverage of data structure requirements
-
-Complex orchestration services differ from simple domain services by:
-- Coordinating multiple external services and dependencies
-- Managing complex multi-step workflows with caching and validation
-- Handling sophisticated data transformations and validations
-- Providing high-level business logic interfaces for use cases
-- Managing performance optimizations like intelligent caching
-- Coordinating security validation across multiple service boundaries
-
-Testing philosophy:
-- Test the contract, not the implementation
-- Validate data integrity and structure completeness
-- Ensure architectural patterns are properly followed
-- Focus on business logic interfaces and capabilities
-- Maintain fast, reliable tests that don't break with refactoring
-*/
